// src/types/ActionTypes.ts
export enum AppActionTypes {
  TOGGLE_GUESS = 'TOGGLE_GUESS',
  ADD_GUESS = 'ADD_GUESS',
  REMOVE_GUESS = 'REMOVE_GUESS',
  NEXT_QUIZ = 'NEXT_QUIZ',
  RESTART_GAME = 'RESTART_GAME',
  SELECT_GAME_MODE = 'SELECT_GAME_MODE',
  CLEAR_GAME_MODE = 'CLEAR_GAME_MODE',
  REDUCE_TRIES = 'REDUCE_TRIES'
}
export enum TimerProgressActionTypes {
  SET_PROGRESS = 'SET_PROGRESS'
}